Thermo.java used to perform the functions that the code ThermoDataEstimator.java now performs in RMG 3.2. This code has been modified to output a bunch of thermochemical properties of a species (in the solution phase) and is used mainly as a debugging tool to check for bugs in the group additivity schemes for solvation, output solvation related properties of a set of species for comparisons, etc.
The current version of this code outputs the Abraham parameters
(A, B, E, L, S, V) of the species and the radius calculated using the
UNIFAC group additivity method. The input file to this code requires
specification of the solvation flag and the temperature followed by the
chemgraphs of the species whose properties are desired.