Updated rates rules for R. + O2 = HO2 + alkene
RMG was overestimating the disproportionation reaction
between R. and O2 leading to severe problems in alkane
oxidation models. The corresponding rate rules have
been corrected using recently reported values by
Miyoshi (dx.doi.org/10.1021/jp112152n). It should be
noted that the flux of this reaction is weak at low
temperatures which was not the case with the previous
rate rule which assumed a barrierless reaction.