Allow O2 as a valid species if MAX_RAD_NUM = 1
This is rather temporary fix to allowing oxygen as a valid reactant
if the user wants to restrict the maximum number of radicals per
species to 1. Currently, for oxidation mechanisms, setting the maximum
allowed number of radical centers to 1 causes RMG to reject O2 as a
valid reactant. I have seen messages like "This species is forbidden but
its in a thermo library..blah blah... so we will allow it" after which
RMG crashes so I am not sure what went wrong there. I have had to remove
some birads from the primary thermo library to get this working.